requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pypi_0
blas=1.0=mkl
bottleneck=1.3.5=py39h7deecbd_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.7=pypi_0
cloudpickle=1.2.2=pypi_0
community=1.0.0b1=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.11.0=pypi_0
filelock=3.8.0=pypi_0
flask=2.2.2=pypi_0
fonttools=4.42.1=pypi_0
future=0.18.2=pypi_0
gym=0.15.3=pypi_0
gym-notices=0.0.8=pypi_0
idna=3.4=pypi_0
importlib-metadata=5.0.0=pypi_0
importlib-resources=6.1.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
markupsafe=2.1.1=pypi_0
matplotlib=3.8.0=pypi_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
ncurses=6.4=h6a678d5_0
networkx=2.8.6=pypi_0
numexpr=2.8.4=py39hc78ab66_1
numpy=1.23.2=pypi_0
numpy-base=1.25.2=py39hb5e798b_0
openssl=3.0.11=h7f8727e_2
ortools=9.4.1874=pypi_0
packaging=23.1=pypi_0
pandas=2.0.3=py39h1128e8f_0
pfrl=0.3.0=pypi_0
pillow=10.0.1=pypi_0
pip=23.2.1=py39h06a4308_0
protobuf=4.21.5=pypi_0
pyglet=1.3.2=pypi_0
pyparsing=3.1.1=pypi_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-louvain=0.16=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3.post1=py39h06a4308_0
readline=8.2=h5eee18b_0
requests=2.28.1=pypi_0
scikit-learn=1.1.3=pypi_0
scipy=1.9.3=pypi_0
setuptools=68.0.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
torch=1.7.1=pypi_0
tqdm=4.65.0=py39hb070fc8_0
typing-extensions=4.3.0=pypi_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.12=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.38.4=py39h06a4308_0
xz=5.4.2=h5eee18b_0
zipp=3.9.0=pypi_0
zlib=1.2.13=h5eee18b_0
